Added SchedulerController.kt: add controller to configure scheduller
Edited Koin.kt:
added SchedulerController.kt

Edited Routing.kt:
now it has no business logic
