+++ src/main/kotlin/team/microchad/plugins/Routing.kt
@@ -10,15 +10,13 @@ import org.koin.ktor.ext.inject
 import team.microchad.client.JiraClient
 import team.microchad.client.MmClient
 import team.microchad.controllers.RegistrationController
+import team.microchad.controllers.SchedulerController
 import team.microchad.controllers.StatisticsController
 import team.microchad.dto.mm.IncomingMsg
-import team.microchad.dto.mm.OutgoingMsg
 import team.microchad.dto.mm.dialog.Response
 import team.microchad.dto.mm.dialog.submissions.SchedulerSubmission
 import team.microchad.dto.mm.dialog.submissions.SelectionSubmission
 import team.microchad.dto.mm.dialog.submissions.StatisticsSubmission
-import team.microchad.dto.mm.slash.*
-import team.microchad.model.entities.ProjectMap
 import team.microchad.model.repositories.ProjectMapRepository
 import team.microchad.service.*
 
@@ -31,6 +29,7 @@ fun Application.configureRouting() {
 
     val registrationController: RegistrationController by inject()
     val statisticsController: StatisticsController by inject()
+    val schedulerController: SchedulerController by inject()
 
     routing {
 
@@ -50,11 +49,9 @@ fun Application.configureRouting() {
             call.respond(actionResponse)
         }
         post("/scheduler_dialog") {
-            val projects = jiraClient.getProjects()
             val incomingMsg = call.receive<IncomingMsg>()
-            val dialog = createSchedulerDialog(incomingMsg.triggerId, projects)
-            mmClient.openDialog(dialog)
-            call.respond(ActionResponse("Continue setting in dialog window"))
+            val actionResponse = schedulerController.openDialog(incomingMsg)
+            call.respond(actionResponse)
         }
 
         post("/statistics_dialog") {
@@ -75,17 +72,7 @@ fun Application.configureRouting() {
 
         post("/scheduler") {
             val incoming = call.receive<Response<SchedulerSubmission>>()
-            val project = incoming.submission?.selectProject?:""
-            val radioScheduler = incoming.submission?.radioScheduler
-            val channelId = incoming.channelId
-            val projectMap = ProjectMap(null, project, channelId)
-            if (radioScheduler == "on") {
-                projectMapRepository.create(projectMap)
-            } else {
-                //TODO DELETE FROM DB
-            }
-            val outgoingMessage = OutgoingMsg(channelId, "$project, $radioScheduler, ${incoming.channelId}")
-            mmClient.sendToDirectChannel(outgoingMessage)
+            schedulerController.configureScheduler(incoming)
         }
 
         post("/register_user") {
